Removed unreferenced variable "rnd"
The variable `rnd` was declared and set but then is not used on line 755. Instead, `rounds` is used, which causes the function to error if a `rounds` argument is not provided and instead the function is called with only a callback argument. This causes an asynchronous exception, `Error: Illegal argument types: function` to be thrown on line 756. The code now mirrors that of the `genSaltSync` method.