Generate 2b hashes by default
BREAKING CHANGE: This library was not affected by the bug that led to
incrementing the bcrypt version from 2a to 2b, but nowadays most
implementations use 2b, including the native bcrypt binding, so this
change aligns with them. Existing hashes will continue to work, but
test logic that generates hashes and compares them literally might need
to be updated to account for the new default.